+++ index.html
@@ -248,6 +248,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 </body>
-<script src="js/main.js"></script>
+<script src="js/main.js" type="module"></script>
 
 </html>
